feat: improve ai component structure and performance
