+++ lib/ai/types.ts
@@ -1,0 +1,5 @@
+import type { InferUITools, UIMessage } from "ai"
+import type { tools } from "@/lib/ai/tools"
+
+// The UIMessage type of the NovaCrate AI Assistant
+export type NC_UIMessage = UIMessage<never, never, InferUITools<typeof tools>>
